Add nlp namespace
In order to get more consistent results from the Recipe Puppy API, I
added the ability to extract nouns from a string. The thought is that
this will map more closely to ingredients that the API recognizes. And
it was a find excuse for getting my feet wet with NLP.